Transpose of Matrix
You are given a square matrix of size n x m. where n is the number of rows and m is the number of columns.  Your task is to find the transpose of the given matrix.
The transpose of a matrix is obtained by converting all the rows to columns and all the columns to rows.